refactor: move individual milestones into a separate file
